put detailed in extra file
